+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Fix return_of_invalid_type analysis error for ObjCBlocks.
 - Fix crash in ObjC methods and blocks that return structs by value.
 - Fix ObjC methods returning instancetype having the wrong type in sublasses.
+- When generating typedefs for `Pointer<NativeFunction<Function>>`, also
+  generate a typedef for the `Function`.
 - Bump min SDK version to 3.2.0-114.0.dev.
 
 # 9.0.1